Mannys First Commit
I have added the Panel and all of its buttons are under its own hirearchy so they can be accessed.

The first 2 buttons on the top-left of the screen have the script that I am currently working on to add more immersion. So far they change the material color on the cube that will in the future power a light.